dateAgo 함수 따로 빼서 작성 및 수정
